Only subscribe once on startup

+++ controllers/notifications.js
@@ -7,45 +7,46 @@ const subscriptionRecords = require('../services/subscriptionRecords');
 
 exports.post = function *postNotification() {
   const body = yield parse(this);
-  let subscriptionId = body.id;
-  let correspondingSourceTransaction =
-    subscriptionRecords.get(subscriptionId);
-  if (body.resource.state === 'completed' && correspondingSourceTransaction) {
+  if (body.event === 'transfer.update') {
+    let transferId = body.resource.id;
+    let correspondingSourceTransaction = subscriptionRecords.get(transferId);
+    if (body.resource.state === 'completed' && correspondingSourceTransaction) {
 
-    log.debug('got notification about completed destination_transfer');
+      log.debug('got notification about completed destination_transfer');
 
-    // Take execution_condition_fulfillment from the source transaction
-    // that has been executed and send it to the source ledger
-    // to unlock the money that's been held for us
-    correspondingSourceTransaction.execution_condition_fulfillment =
-      body.resource.execution_condition_fulfillment;
+      // Take execution_condition_fulfillment from the source transaction
+      // that has been executed and send it to the source ledger
+      // to unlock the money that's been held for us
+      correspondingSourceTransaction.execution_condition_fulfillment =
+        body.resource.execution_condition_fulfillment;
 
-    let sourceTransactionReq = yield request({
-      method: 'put',
-      url: correspondingSourceTransaction.id,
-      json: true,
-      body: correspondingSourceTransaction
-    });
+      let sourceTransactionReq = yield request({
+        method: 'put',
+        url: correspondingSourceTransaction.id,
+        json: true,
+        body: correspondingSourceTransaction
+      });
 
-    if (sourceTransactionReq.statusCode >= 400) {
-      // TODO handle this so we actually get our money back
-      log.error('error unlocking funds from source_transfer',
-        sourceTransactionReq.body);
-    } else {
-      log.debug('unlocked source_transfer funds');
+      if (sourceTransactionReq.statusCode >= 400) {
+        // TODO handle this so we actually get our money back
+        log.error('error unlocking funds from source_transfer',
+          sourceTransactionReq.body);
+      } else {
+        log.debug('unlocked source_transfer funds');
+      }
+
+      let removeSubscriptionReq = yield request({
+        method: 'DELETE',
+        uri: subscriptionId
+      });
+      if (removeSubscriptionReq.statusCode >= 400) {
+        log.error('error removing subscription', removeSubscriptionReq.body);
+      } else {
+        log.debug('removed subscription', subscriptionId);
+      }
+
+      subscriptionRecords.remove(subscriptionId);
     }
-
-    let removeSubscriptionReq = yield request({
-      method: 'DELETE',
-      uri: subscriptionId
-    });
-    if (removeSubscriptionReq.statusCode >= 400) {
-      log.error('error removing subscription', removeSubscriptionReq.body);
-    } else {
-      log.debug('removed subscription', subscriptionId);
-    }
-
-    subscriptionRecords.remove(subscriptionId);
   }
 
   this.status = 200;
